test(singup api): tests for PATCH and DELETE methods

+++ src/app/api/users/route.ts
@@ -1,5 +1,6 @@
 import prisma from '@/lib/db'
 import { userSchema } from '@/schemas/user'
+import { excludeFromList, excludeFromObject } from '@/utils/ExcludeResponseParam'
 import { hashPass } from '@/utils/HashManager'
 import { NextRequest, NextResponse } from 'next/server'
 
@@ -18,7 +19,7 @@ export async function GET() {
   try {
     const users = await prisma.users.findMany()
 
-    return NextResponse.json(users, { status: 200 })
+    return NextResponse.json(excludeFromList(users, ['password']), { status: 200 })
   } catch (error) {
     return NextResponse.json({ message: 'Error in fetching users' + error }, { status: 500 })
   }
@@ -30,7 +31,7 @@ export async function POST(req: NextRequest) {
     const response = userSchema.parse(body)
     const { firstName, lastName, email, password } = response
 
-    const findUser = await prisma.users.findFirst({ where: { email } })
+    const findUser = await prisma.users.findUnique({ where: { email } })
 
     if (findUser) {
       return NextResponse.json({ message: 'E-mail already registered' }, { status: 409 })
@@ -50,7 +51,7 @@ export async function POST(req: NextRequest) {
     return NextResponse.json(
       {
         message: 'User created successfully',
-        user: newUser,
+        user: excludeFromObject(newUser, ['password']),
       },
       { status: 201 },
     )
@@ -59,8 +60,14 @@ export async function POST(req: NextRequest) {
   }
 }
 
-export async function PATCH(req: Request) {
+export async function PATCH(req: NextRequest) {
   const { id, email, password } = await req.json()
+
+  const findUser = await prisma.users.findUnique({ where: { id } })
+
+  if (!findUser) {
+    return NextResponse.json({ message: 'User not found' }, { status: 404 })
+  }
 
   const updatedUser = await prisma.users.update({
     where: { id },
@@ -70,7 +77,7 @@ export async function PATCH(req: Request) {
   return NextResponse.json(
     {
       message: 'User updated successfully',
-      user: updatedUser,
+      user: excludeFromObject(updatedUser, ['password']),
     },
     { status: 201 },
   )
@@ -79,6 +86,12 @@ export async function PATCH(req: Request) {
 export async function DELETE(req: Request) {
   const { id } = await req.json()
 
+  const findUser = await prisma.users.findUnique({ where: { id } })
+
+  if (!findUser) {
+    return NextResponse.json({ message: 'User not found' }, { status: 404 })
+  }
+
   const deletedUser = await prisma.users.delete({
     where: { id },
   })
@@ -86,7 +99,7 @@ export async function DELETE(req: Request) {
   return NextResponse.json(
     {
       message: 'User deleted successfully',
-      user: deletedUser,
+      user: excludeFromObject(deletedUser, ['password']),
     },
     { status: 201 },
   )
